Import usbwcm driver and headers from illumos-gate
As described in the README, this repository is a downstream "fork" of
just the usbwcm STREAMS module that is maintained within the illumos-gate
repository at https://github.com/illumos/illumos-gate. The purpose of
this repository is to serve as a convenient way for users of Illumos,
Solaris, etc. to broaden the range of Wacom tablets that their system
can use without requiring an upgrade to the entire kernel.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>